extract memory monitor into separate package
